+++ collector/processor/concurrentbatchprocessor/generated_package_test.go
@@ -3,9 +3,8 @@
 package concurrentbatchprocessor
 
 import (
-	"testing"
-
 	"go.uber.org/goleak"
+	"testing"
 )
 
 func TestMain(m *testing.M) {